services/indexer/requirements.txt
requests>=2.31.0
opensearch-py>=2.3.0
pyyaml>=6.0
python-consul>=1.1.0
python-dotenv>=1.0.0
